Sorting of folder, renaming of scripts, movement speed fixes
New Features:
None.
Changes:
Relocated folders to create a better folder structure.
Renamed some scripts to reflect the majority of what the script does.
Increased the drag on the player to make it look as if they stop immediately. Requires further testing to get the right balance for the game.