Addition of some animations, as well as a beta of the new food assets
Additions:
Replaced the generic ball objects for burger sprites
Added animations for the new powerups as a glowing effect
Bug Fixes/Updates:
Renamed some elements within scripting
Added the chance of powerups spawning to the spawning objects script, as well as a breakdown in comment form of the RNG elements
Updated the text for powerups so that both can now be displayed at the same time.
Bugs found:
Not all animations play correctly.