Added floor asset, as well as a major bug fix.
Additions:
Downloaded a floor asset and implemented. Based on feedback this may be updated.
Bug fixes/Updates:
All of the spawns from the dumpsters now spawn in a random direction, instead of not moving and just rolling to the floor.

Updated the health and score script so that the threshold score and health can be changed in the editor