New map for level 1, patrolling enemy implemented including damage counter, new raycasting script, new animations
Additions:
Patrolling enemies script has been added to include more points as well as damaging the player. This will be implemented in further levels, except the tutorial level.
New text has been added for warning the player they are about to be damaged.
New animation for the text, flashes when it is on the screen so that the player will not miss it.
Extra cubes have been added below the map to be used as a pathfinding system for the raccoon.

Updates:
The level 1 map has been updated. The icon in the level select scene will need to be updated next.

Next steps:
Rebalancing level 1, as currently the player does not have enough time.
Updating the map icon on the level select screen to reflect the new addition.