Level 1 tweaks, sound passing through scenes fix
Additions:
None
Updates:
Rebalanced level 1 a little bit:
Enemy now has a bigger spherecast radius, so the player is more likely to be damaged, as they were no threat originally.
The number of points required to complete the first level has been increased from 7 to 10.
The amount of health the player has in level 1 has been reduced from 5 to 3.
This is still subject to balancing - although I think it is about what I would like for the first level.

Next steps:
Preparation for demo - including the addition of a reset button for the best time, as it may be inconsistent with the new difficulty.
Add animations/sounds, and create floor plans for levels 2, 3 and 4.
Modularise code.